refactor: split models in files
