+++ raspberrypi_central/webapp/app/alarm/admin.py
@@ -1,8 +1,11 @@
 from django.contrib import admin
+
+import alarm.models.alarm
+import alarm.models.camera
 from . import models
 
-admin.site.register(models.AlarmStatus)
-admin.site.register(models.AlarmSchedule)
-admin.site.register(models.CameraMotionDetected)
-admin.site.register(models.CameraMotionDetectedPicture)
-admin.site.register(models.CameraRectangleROI)
+admin.site.register(alarm.models.alarm.AlarmStatus)
+admin.site.register(alarm.models.alarm.AlarmSchedule)
+admin.site.register(alarm.models.camera.CameraMotionDetected)
+admin.site.register(alarm.models.camera.CameraMotionDetectedPicture)
+admin.site.register(alarm.models.camera.CameraRectangleROI)
