+++ raspberrypi_central/webapp/app/alarm/external/motion.py
@@ -2,6 +2,7 @@ from typing import List, Dict
 
 from django.db import IntegrityError
 
+import alarm.models.camera
 from alarm import models as alarm_models
 from devices import models as device_models
 
@@ -10,7 +11,7 @@ def save_motion(device_id: str, seen_in: Dict[str, List[str]], event_ref: str, s
     device = device_models.Device.objects.get(device_id=device_id)
 
     try:
-        motion = alarm_models.CameraMotionDetected.objects.create(device=device, event_ref=event_ref, is_motion=status)
+        motion = alarm.models.camera.CameraMotionDetected.objects.create(device=device, event_ref=event_ref, is_motion=status)
         motion.save()
     except IntegrityError:
         return None, None
